requirement.txt
python-socketio[client]=4.6.0
python-socketio==4.6.0
